+++ yaqd_attune/_attune.py
@@ -49,10 +49,7 @@ class Attune(HasDependents, HasLimits, IsHomeable, HasPosition, IsDaemon):
         for name, set_pos in self._instrument(position, self._state["arrangement"]).items():
             if name in exceptions:
                 continue
-            try:
-                set_pos = set_pos.item()
-            except AttributeError:
-                pass
+            set_pos = set_pos.item()  # needs attune > 0.5.1
             if isinstance(set_pos, str):
                 self._setables[name].set_identifier(set_pos)
             elif set_pos is None: