+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [Unreleased]
 
 - attune now works up to python 3.11 (resolved upstream)
+- integration to work with attune 0.5.2
 
 ## [2023.3.0]